pyproject.toml
[tool.poetry]
name = "python_django_diploma"
version = "0.1.0"
description = ""
authors = ["andrei <mezentsev0904@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.2"
django-rest-framework = "^0.1.0"
gunicorn = "^20.1.0"
Pillow = "^9.5.0"
python-dotenv = "^1.0.0"
pytz = "^2023.3"
requests = "^2.31.0"
sentry-sdk = "^1.26.0"
charset-normalizer = "^3.1.0"
typing-extensions = "^4.6.3"
urllib3 = "^2.0.3"
diploma-frontend = {path = "diploma-frontend-0.6.tar.gz"}
whitenoise = "^6.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
